moved test utility functions
